Refacted the utils to use separate files, and ES7 export syntaxes
